+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## UNRELEASED
+
+### Adds
+
+* Add `_edit: true` to the list of locales the current user can edit in `@apostrophecms/i18n`.
+
 ## 3.42.0 (2023-03-16)
 
 ### Adds